Create insert-lang-attribute template for jats2opus4 (including mapping language codes from ISO 639-1 to 639-2), updated jats2metsmods's insert-lang-attribute template to cover higher-level elements
For jats2opus4:
I added an "insert-lang-attribute" template which looks for the current element's language attribute and alternatively looks for the next higher-level element's language attribute. That language code is then mapped to the three-character ISO code that OPUS4 requires. If no language tag is found in either element, the document language $langOut gets used.
template is currently applied for translated titles, abstracts, and for keywords.

for jats2metsmods: expanded this insert-lang-attribute template to also look for language attributes in the higher-level element.

language attributes are located in higher level elements in case of translated titles, and multi-language keywords.